ml-service/scraper.py
import requests
from bs4 import BeautifulSoup
import time
from typing import List, Dict
import random

class PropertyScraper:
    def __init__(self):
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        }
        self.delay = 2  # Seconds between requests
    
    async def scrape_all_sources(self, district: str, state: str) -> List[Dict]:
        """Scrape all available sources"""
        results = []
        
        # Note: Actual scraping would require proper implementation
        # For now, returning sample data structure
        print(f"Scraping data for {district}, {state}")
        
        return results
    
    async def get_market_data(self, district: str, state: str, land_type: str) -> Dict:
        """Get market insights for a location"""
        # This would aggregate scraped data
        # For now, return sample insights
        
        market_insights = {
            'avg_price_per_sqft': self._get_avg_price(district, state),
            'price_trend': 'increasing',
            'growth_rate': 8.5,  # % per year
            'recent_sales': 15,
            'market_activity': 'moderate',
            'comparable_properties': []
        }
        
        return market_insights
    
    def _get_avg_price(self, district: str, state: str) -> float:
        """Get average price for area"""
        # Sample data - would come from database
        prices = {
            'Kerala': {
                'Ernakulam': 2800,
                'Thiruvananthapuram': 2500,
                'Kozhikode': 2200,
                'Thrissur': 2100,
                'Kannur': 1900
            },
            'Karnataka': {
                'Bangalore': 4500,
                'Mysore': 2000,
                'Mangalore': 2300
            },
            'Tamil Nadu': {
                'Chennai': 4000,
                'Coimbatore': 2200,
                'Madurai': 1800
            },
            'Maharashtra': {
                'Mumbai': 8000,
                'Pune': 3500
            }
        }
        
        return prices.get(state, {}).get(district, 2000)
    
    async def scrape_magicbricks(self, location: str) -> List[Dict]:
        """Scrape MagicBricks (placeholder - would need actual implementation)"""
        # Actual implementation would scrape MagicBricks
        # This is a placeholder showing the structure
        return []
    
    async def scrape_99acres(self, location: str) -> List[Dict]:
        """Scrape 99acres (placeholder)"""
        return []
