ml-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
requests==2.31.0
beautifulsoup4==4.12.3
googlemaps==4.10.0
python-dotenv==1.0.0
pydantic==2.5.3
joblib==1.3.2
